+++ zetaclient/chains/bitcoin/observer/inbound_tracker.go
@@ -3,6 +3,7 @@ package observer
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"time"
 
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
@@ -12,6 +13,7 @@ import (
 	"github.com/zeta-chain/node/zetaclient/chains/bitcoin/common"
 	"github.com/zeta-chain/node/zetaclient/config"
 	"github.com/zeta-chain/node/zetaclient/logs"
+	"github.com/zeta-chain/node/zetaclient/metrics"
 	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
 
@@ -52,7 +54,7 @@ func (ob *Observer) observeInboundTrackers(
 			Stringer(logs.FieldCoinType, tracker.CoinType).
 			Bool("is_internal", isInternal).
 			Msg("processing inbound tracker")
-		if _, err := ob.CheckReceiptForBtcTxHash(ctx, tracker.TxHash, true); err != nil {
+		if _, err := ob.CheckReceiptAndPostVoteForBtcTxHash(ctx, tracker.TxHash, true, isInternal); err != nil {
 			return err
 		}
 	}
@@ -60,8 +62,9 @@ func (ob *Observer) observeInboundTrackers(
 	return nil
 }
 
-// CheckReceiptForBtcTxHash checks the receipt for a btc tx hash
-func (ob *Observer) CheckReceiptForBtcTxHash(ctx context.Context, txHash string, vote bool) (string, error) {
+// CheckReceiptAndPostVoteForBtcTxHash checks the receipt for a btc tx hash
+func (ob *Observer) CheckReceiptAndPostVoteForBtcTxHash(ctx context.Context,
+	txHash string, vote bool, isInternal bool) (string, error) {
 	hash, err := chainhash.NewHashFromStr(txHash)
 	if err != nil {
 		return "", errors.Wrap(err, "error parsing btc tx hash")
@@ -125,6 +128,7 @@ func (ob *Observer) CheckReceiptForBtcTxHash(ctx context.Context, txHash string,
 		return msg.Digest(), nil
 	}
 
+	metrics.InboundObservationsTrackerTotal.WithLabelValues(ob.Chain().Name, strconv.FormatBool(isInternal)).Inc()
 	return ob.ZetaRepo().VoteInbound(ctx,
 		ob.logger.Inbound,
 		msg,